llm_assistant/requirements.txt
fastapi
uvicorn
requests
phidata
bs4
sqlalchemy
pgvector
psycopg[binary]
pypdf
ollama
httpx